move TagVerifier.java
