+++ plugins/httpservlet/src/test/java/com/megaease/easeagent/plugin/httpservlet/interceptor/DoFilterMetricInterceptorTest.java
@@ -21,7 +21,7 @@ import com.megaease.easeagent.mock.metrics.MockMetricUtils;
 import com.megaease.easeagent.mock.plugin.api.MockEaseAgent;
 import com.megaease.easeagent.mock.report.ReportMock;
 import com.megaease.easeagent.mock.report.impl.LastJsonReporter;
-import com.megaease.easeagent.mock.utils.TagVerifier;
+import com.megaease.easeagent.mock.plugin.api.utils.TagVerifier;
 import com.megaease.easeagent.plugin.api.config.IPluginConfig;
 import com.megaease.easeagent.plugin.api.metric.ServiceMetric;
 import com.megaease.easeagent.plugin.bridge.EaseAgent;
